requirements.txt
appdirs==1.4.3
Django==3.0.5
django-enumchoicefield==2.0.0
djangorestframework==3.11.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1

